部分完善动态查看
年前写的没写完。。。

+++ bili-api/src/main/kotlin/dev/aaa1115910/biliapi/http/BiliHttpApi.kt
@@ -1526,7 +1526,7 @@ object BiliHttpApi {
      */
     suspend fun getComments(
         type: Int,
-        oid: Int,
+        oid: Long,
         mode: Int = 3,
         paginationStr: String = """{"offset":""}""",
         //webLocation: Int = 1815875,
@@ -1543,7 +1543,7 @@ object BiliHttpApi {
         }.body()
 
     suspend fun getCommentReplies(
-        oid: Int,
+        oid: Long,
         type: Int,
         root: Long,
         pageSize: Int = 10,
